# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PhaseOne10841 Agent Security Gateway API
  description: |
    Defensive Agent EDR / security gateway by Veracity Integrity LLC.
    https://VeracityIntegrity.com

    Covers the gateway (OpenAI-compatible proxy + PhaseOne control APIs)
    and notes on the MFA admin dashboard proxy surface.
  version: 0.5.0
  contact:
    name: Veracity Integrity LLC
    url: https://VeracityIntegrity.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local gateway (docker compose)
  - url: http://localhost:3000
    description: Admin dashboard (proxies /api/* → gateway)
...
paths:
...
  /v1/chat/completions:
    post:
      tags: [Chat]
      summary: OpenAI-compatible chat completions (policy + scan + proxy)
      parameters:
        - in: header
          name: X-PhaseOne-Agent-Id
          schema: { type: string }
        - in: header
          name: X-PhaseOne-Session-Id
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model: { type: string }
                messages:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Upstream or mock completion
        '403':
          description: Blocked by policy / injection / egress
        '429':
          description: Rate limited
...
  /v1/phaseone/events:
    get:
      tags: [Admin]
      summary: List recorded events
      parameters:
        - in: query
          name: session_id
          schema: { type: string }
        - in: query
          name: severity
          schema: { type: string }
        - in: query
          name: event_type
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        '200': { description: Event list }
    post:
      tags: [Admin]
      summary: Record an observable hook event
      responses:
        '201': { description: Created }
...
  /v1/phaseone/retention/run:
    post:
      tags: [Ops]
      summary: Run retention cleanup (defaults to dry_run=true via API)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                days: { type: integer }
                dry_run: { type: boolean, default: true }
      responses:
        '200': { description: Result }
...
  /api/auth/request-otp:
    post:
      tags: [DashboardAuth]
      summary: Request MFA OTP (dashboard, rate limited)
      servers:
        - url: http://localhost:3000
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
      responses:
        '200': { description: Sent / logged }
        '429': { description: OTP rate limited }
...